Move sliceviewer views to subdirectory
