+++ qt/python/mantidqt/mantidqt/widgets/sliceviewer/views/cursor.py
