+++ qt/python/mantidqt/mantidqt/widgets/sliceviewer/views/dimensionwidget.py
