+++ qt/python/mantidqt/mantidqt/widgets/sliceviewer/views/toolbar.py
